backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
python-dotenv
groq
requests
reportlab
qrcode
pillow
scikit-learn
xgboost
joblib